fix: Fix `AL` comments. - Extract the assets dropdown on separate component.
